requirements.txt
streamlit
PyPDF2 
pandas 
beautifulsoup4 
trafilatura 
validators 
requests
chromadb
ragas 
datasets 
plotly
langgraph 
langchain 
langchain-google-genai 
langchain-community
sentence-transformers
langchain-huggingface
langchain-anthropic 
langchain-openai
langchain-ollama
langchain-groq
ollama
einops
python-dotenv
flashrank
rank_bm25
